opendata/requirements.txt
django==2.1.1
markdown
django-filter
djangorestframework
django-rest-swagger
PyYaml
